+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Two menu items have shorter names: "Syncthing Menu Settings…" is now "Settings…", and "Open Syncthing Web UI" is now "Open Syncthing".
+
 ## [0.1.7] - 2026-07-09
 
 ### Changed